+++ src/job-parse.js
@@ -1,6 +1,6 @@
 const logger = require('./logger')(module)
 const is = require('./is')
-const moment = require('moment')
+const datetime = require('./datetime')
 const enums = require('./enums')
 
 module.exports.id = function jobParseId (job) {