+++ src/logger.js
@@ -1,9 +1,9 @@
 const path = require('path')
-const moment = require('moment')
+const datetime = require('./datetime')
 const debug = require('debug')
 module.exports = function logger (rjqModule) {
   if (process.env.DEBUG) {
-    const time = moment().format('HH:mm:ss.SSS')
+    const time = datetime.format(new Date())
     const moduleName = path.basename(rjqModule.id, '.js')
     let prefix = `[${time}][${moduleName}]`
     return debug(prefix)